fixes, move apis to hub
