+++ api/hubd/pb/hubd.proto
@@ -178,6 +178,71 @@ message DestroyAccountRequest {}
 
 message DestroyAccountResponse {}
 
+message ArchivesLsRequest {
+}
+
+message ArchivesLsResponse {
+    repeated ArchiveLsItem archives = 1;
+}
+
+message ArchiveLsItem {
+    string cid = 1;
+    repeated ArchiveLsItemMetadata info = 2;
+}
+
+message ArchiveLsItemMetadata {
+    uint64 deal_id = 1;
+}
+
+message ArchivesImportRequest {
+    string cid = 1;
+    repeated uint64 deal_ids = 2;
+}
+
+message ArchivesImportResponse {}
+
+message ArchiveRetrievalLsRequest {}
+
+message ArchiveRetrievalLsResponse {
+    repeated ArchiveRetrievalLsItem retrievals = 1;
+}
+
+message ArchiveRetrievalLsItem {
+	string id = 1;
+	string cid = 2;
+	ArchiveRetrievalStatus status = 3;
+	string FailureCause = 4;
+	int64 created_at = 5;
+
+	oneof retrieval_type {
+		ArchiveRetrievalLsItemNewBucket newBucket = 6;
+	}
+}
+
+message ArchiveRetrievalLsItemNewBucket {
+	string name = 1;
+	bool private = 2;
+}
+
+enum ArchiveRetrievalStatus {
+    UNSPECIFIED = 0;
+    QUEUED = 1;
+    EXECUTING = 2;
+    MOVETOBUCKET = 3;
+    FAILED = 4;
+    CANCELED = 5;
+    SUCCESS = 6;
+}
+
+message ArchiveRetrievalLogsRequest {
+    string id = 1;
+}
+
+message ArchiveRetrievalLogsResponse {
+    string msg = 1;
+}
+
+
 service APIService {
     rpc BuildInfo(BuildInfoRequest) returns (BuildInfoResponse) {}
 
@@ -207,4 +272,12 @@ service APIService {
     rpc IsOrgNameAvailable(IsOrgNameAvailableRequest) returns (IsOrgNameAvailableResponse) {}
 
     rpc DestroyAccount(DestroyAccountRequest) returns (DestroyAccountResponse) {}
+
+    // Archives Import
+    rpc ArchivesLs(ArchivesLsRequest) returns (ArchivesLsResponse) {}
+    rpc ArchivesImport(ArchivesImportRequest) returns (ArchivesImportResponse) {}
+
+    // Archive Retrieval
+    rpc ArchiveRetrievalLs(ArchiveRetrievalLsRequest) returns (ArchiveRetrievalLsResponse){}
+    rpc ArchiveRetrievalLogs(ArchiveRetrievalLogsRequest) returns (stream ArchiveRetrievalLogsResponse) {}
 }
